Pixel classifier add resolution and static methods
Implemented 'Add resolution' option, extracted static methods to apply classifications to detections and to create objects from classification.
